+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add `Pay::Merchant` and Stripe Connect functionality - @excid3
 * Save `currency` on Pay::Charge records - @excid3
 
 ### 2.6.11