+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Refactor Stripe webhooks to always retrieve latest records - @excid3
 * Add `Pay::Merchant` and Stripe Connect functionality - @excid3
 * Save `currency` on Pay::Charge records - @excid3